requirements.txt
pytest
pytest-cov
Pillow
numpy
scikit-image
opencv-python
hypothesis
pytest-benchmark
pytest-mock
imagehash
